Set up Kotlin Multiplatform tests on JVM
