+++ examples/kotlin/.gitignore
@@ -9,3 +9,6 @@
 .externalNativeBuild
 .cxx
 local.properties
+mongodb-realm
+*.realm
+*.management
