--- examples/kotlin/shared/src/commonTest/kotlin/com/mongodb/realm/realmkmmapp/syncTest.kt
@@ -1,4 +1,0 @@
-package com.mongodb.realm.realmkmmapp
-
-class SyncTest {
-}
