+++ examples/kotlin/shared/src/jvmMain/kotlin/com/mongodb/realm/realmkmmapp/Platform.kt
@@ -1,0 +1,5 @@
+package com.mongodb.realm.realmkmmapp
+
+actual class Platform actual constructor() {
+    actual val platform: String = "JVM ${System.getProperty("java.version")}"
+}
