+++ source/sdk/kotlin/app-services/authenticate-users.txt
@@ -53,7 +53,7 @@ authentication, create an anonymous credential by calling
 and then pass the generated credential to
 ``app.login()``.
 
-.. literalinclude:: /examples/generated/kotlin/Greeting.codeblock.anonymous-authentication.kt
+.. literalinclude:: /examples/generated/kotlin/AuthenticationTest.codeblock.anonymous-authentication.kt
    :language: kotlin
 
 .. _kotlin-email-password-login:
@@ -70,7 +70,7 @@ email/password credential by calling :kotlin-sdk:`AuthenticationProvider.EMAIL_P
 with the user's email and password. Then pass the generated credential
 to ``app.login()``.
 
-.. literalinclude:: /examples/generated/kotlin/Greeting.codeblock.email-password-authentication.kt
+.. literalinclude:: /examples/generated/kotlin/AuthenticationTest.codeblock.email-password-authentication.kt
    :language: kotlin
 
 .. _kotlin-logout:
@@ -89,5 +89,5 @@ to log in, using :kotlin-sdk:`user.logOut()
 Because logging out halts synchronization, you should only log out after
 all local Realm updates have uploaded to the server.
 
-.. literalinclude:: /examples/generated/kotlin/Greeting.codeblock.log-out.kt
+.. literalinclude:: /examples/generated/kotlin/AuthenticationTest.codeblock.log-out.kt
    :language: kotlin
