+++ source/sdk/kotlin/app-services/connect-to-a-realm-app.txt
@@ -25,7 +25,7 @@ Access the App Client
 Pass the App ID for your {+app+}, which you can :ref:`find in the Realm UI
 <find-your-app-id>`.
 
-.. literalinclude:: /examples/generated/kotlin/Greeting.codeblock.initialize-app-client.kt
+.. literalinclude:: /examples/generated/kotlin/SyncTest.codeblock.initialize-app-client.kt
    :language: kotlin
 
 .. note::
@@ -45,7 +45,7 @@ such as custom timeouts for connections and keys for local encryption,
 use the :kotlin-sdk:`AppConfiguration.Builder
 <library-sync/-realm%20-kotlin%20-multiplatform%20-s-d-k%20-%20-sync/io.realm.mongodb/-app-configuration/-builder/index.html>` to control details of your ``App``:
 
-.. literalinclude:: /examples/generated/kotlin/Greeting.codeblock.configure-app-client.kt
+.. literalinclude:: /examples/generated/kotlin/SyncTest.codeblock.configure-app-client.kt
    :language: kotlin
 
 .. note::
