+++ source/sdk/kotlin/app-services/register-users.txt
@@ -26,5 +26,5 @@ To register a new user, pass a user-provided email and password to
 :kotlin-sdk:`app.emailPasswordAuth.registerUser()
 <library-sync/-realm%20-kotlin%20-multiplatform%20-s-d-k%20-%20-sync/io.realm.mongodb/-email-password-auth/register-user.html>`:
 
-.. literalinclude:: /examples/generated/kotlin/Greeting.codeblock.email-password-registration.kt
+.. literalinclude:: /examples/generated/kotlin/RegistrationTest.codeblock.email-password-registration.kt
    :language: kotlin
